LPS-54043 Move public API into kernel package
